requirements.txt
Flask==3.1.1
flask-cors==6.0.0
pymongo==4.11.1
python-dotenv==1.1.0
Faker==37.3.0
dnspython==2.6.1
